fix(monkey): cast bracket-commit params to numeric to resolve "operator is not unique"
Postgres parses positional parameters as `unknown` type. The bracket-commit
UPDATE in claimAdoptedPositions had `entry_price + (CASE ... -$2 ...)`, where
unary `-` on `unknown` is ambiguous (operator overload resolution can't pick a
candidate). Result: every tick logged `ERROR: operator is not unique: - unknown
at character 140` on the postgres side and `[Monkey] claimAdoptedPositions
failed {err: "operator is not unique: - unknown"}` on the app side. Soft-fails
correctly per existing test, but silently leaves operator-adopted positions
without TP/SL brackets — synthetic exit gate then has nothing to enforce.

Fix: explicit ::numeric casts on $2 and $3 (entry_price is DECIMAL(30,18) =
NUMERIC). Test contract unchanged.

+++ apps/api/src/services/monkey/loop.ts
@@ -5199,12 +5199,16 @@ export class MonkeyKernel extends EventEmitter {
       // Commit a geometry-derived bracket on any owned row that lacks
       // one — side-aware TP/SL around the row's recorded entry price.
       if (frBracket && frBracket.tpDistance > 0 && frBracket.slDistance > 0) {
+        // Positional params arrive as Postgres type `unknown`; unary
+        // minus on `unknown` is ambiguous ("operator is not unique:
+        // - unknown"). Explicit ::numeric casts resolve it — entry_price
+        // is DECIMAL(30,18) = NUMERIC, so the arithmetic types match.
         await pool.query(
           `UPDATE autonomous_trades
               SET take_profit = entry_price + (CASE
-                    WHEN side IN ('long', 'buy') THEN $2 ELSE -$2 END),
+                    WHEN side IN ('long', 'buy') THEN $2::numeric ELSE -($2::numeric) END),
                   stop_loss   = entry_price + (CASE
-                    WHEN side IN ('long', 'buy') THEN -$3 ELSE $3 END)
+                    WHEN side IN ('long', 'buy') THEN -($3::numeric) ELSE $3::numeric END)
             WHERE reason LIKE 'monkey|kernel=monkey-position|%'
               AND status = 'open' AND symbol = $1
               AND take_profit IS NULL AND stop_loss IS NULL`,
